requirements.txt
# AI Question Generator - Python Dependencies
# Install with: pip install -r requirements.txt

flask>=3.0.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
